Add a box voronoi transform
A "box voronoi" is a variation of a traditional voronoi, where the
voronoi algorithm is run on each dimension separately to return a set
of rectangles instead of a set of irregular polygons.

The algorithm works by binning all items that share the same value in
the first dimension together and creating rectangles for each of those,
then splitting those rectangles using the items in each bin by the
value of the second dimension.