Ensure uglify doesn't convert escaped Unicode sequences
By default, uglify converts escaped Unicode sequences into the Unicode
characters that represent those sequences. This causes problems when
using the minified version with a script tag that doesn't specify a
the correct UTF-8 encoding.

This commit fixes the problem by passing uglify a flag to ensure the
contents of the resulting minified file are ASCII only.